+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # ChangeLog
 
-## v2.0.2 (2024-10-13)
+## v2.0.2 (2024-10-15)
 - Fix Remove Bookmark
-- Add support for store client settings on server
+- Add support for store/restore client settings from server
 
 ## v2.0.1 (2024-10-07)
 - Fix bookmarkTab/permission